Moved error screens into their own errors/ folder
