+++ server/routes/errors/problem-with-service.route.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const { Paths, Views } = require('../utils/constants')
+const { Paths, Views } = require('../../utils/constants')
 
 const handlers = {
   get: (request, h) => {
-    return h.view(Views.PAGE_NOT_FOUND, {
+    return h.view(Views.PROBLEM_WITH_SERVICE, {
       ..._getContext()
     })
   }
@@ -12,14 +12,14 @@ const handlers = {
 
 const _getContext = () => {
   return {
-    pageTitle: 'Page not found'
+    pageTitle: 'Sorry, there is a problem with the service'
   }
 }
 
 module.exports = [
   {
     method: 'GET',
-    path: `${Paths.PAGE_NOT_FOUND}`,
+    path: `${Paths.PROBLEM_WITH_SERVICE}`,
     handler: handlers.get
   }
 ]
