+++ server/utils/constants.js
@@ -77,18 +77,18 @@ const Paths = {
   OWNER_ADDRESS_FIND: '/user-details/owner/address-find',
   OWNER_ADDRESS_INTERNATIONAL: '/user-details/owner/address-international',
   OWNER_CONTACT_DETAILS: '/user-details/owner/contact-details',
-  PAGE_NOT_FOUND: '/page-not-found',
+  PAGE_NOT_FOUND: '/errors/page-not-found',
   PRIVACY_NOTICE: '/privacy-notice',
-  PROBLEM_WITH_SERVICE: '/problem-with-service',
+  PROBLEM_WITH_SERVICE: '/errors/problem-with-service',
   REMOVE_PHOTO: '/remove-photo',
   RMI_AND_PRE_1918: '/eligibility-checker/rmi-and-pre-1918',
   SALE_INTENTION: '/sale-intention',
   SELLING_TO_MUSEUM: '/eligibility-checker/selling-to-museum',
   SERVICE_COMPLETE: '/service-complete',
-  SERVICE_UNAVAILABLE: '/service-unavailable',
+  SERVICE_UNAVAILABLE: '/errors/service-unavailable',
   TAKEN_FROM_ELEPHANT: '/eligibility-checker/taken-from-elephant',
   UPLOAD_PHOTOS: '/upload-photos',
-  UPLOAD_TIMEOUT: '/upload-timeout',
+  UPLOAD_TIMEOUT: '/errors/upload-timeout',
   WHAT_TYPE_OF_ITEM_IS_IT: '/what-type-of-item-is-it',
   WHERE_IS_ITEM: '/where-is-item',
   WHO_OWNS_ITEM: '/who-owns-the-item',
@@ -130,17 +130,17 @@ const Views = {
   LESS_THAN_320CM_SQUARED: 'eligibility-checker/less-than-320cm-squared',
   MADE_BEFORE_1947: 'eligibility-checker/made-before-1947',
   MADE_BEFORE_1975: 'eligibility-checker/made-before-1975',
-  PAGE_NOT_FOUND: 'page-not-found',
+  PAGE_NOT_FOUND: 'errors/page-not-found',
   PRIVACY_NOTICE: 'privacy-notice',
-  PROBLEM_WITH_SERVICE: 'problem-with-service',
+  PROBLEM_WITH_SERVICE: 'errors/problem-with-service',
   RMI_AND_PRE_1918: 'eligibility-checker/rmi-and-pre-1918',
   SALE_INTENTION: 'sale-intention',
   SELLING_TO_MUSEUM: 'eligibility-checker/selling-to-museum',
   SERVICE_COMPLETE: 'service-complete',
-  SERVICE_UNAVAILABLE: 'service-unavailable',
+  SERVICE_UNAVAILABLE: 'errors/service-unavailable',
   TAKEN_FROM_ELEPHANT: 'eligibility-checker/taken-from-elephant',
   UPLOAD_PHOTOS: 'upload-photos',
-  UPLOAD_TIMEOUT: 'upload-timeout',
+  UPLOAD_TIMEOUT: 'errors/upload-timeout',
   WHAT_TYPE_OF_ITEM_IS_IT: 'what-type-of-item-is-it',
   WHERE_IS_ITEM: 'where-is-item',
   WHO_OWNS_ITEM: 'who-owns-the-item',
