+++ test/routes/errors/upload-timeout.route.test.js
@@ -1,12 +1,12 @@
 'use strict'
 
-const createServer = require('../../server')
+const createServer = require('../../../server')
 
-const TestHelper = require('../utils/test-helper')
+const TestHelper = require('../../utils/test-helper')
 
-describe('/upload-timeout route', () => {
+describe('/errors/upload-timeout route', () => {
   let server
-  const url = '/upload-timeout'
+  const url = '/errors/upload-timeout'
   const nextUrl = '/upload-photos'
 
   const elementIds = {
